rename contact inline
